Batch Support in Alarm Processing and increasing historical time range
Daily event jobs is limited to 10000 per org. Assets and specified interval are validated based on the same.
Batch processing of events with more than 5000 count, are handled with autoclear and previous batch lastOccurrence reference. Forthcoming events batch includes both previousBatch lastOccurrence and latestAlarmOccurrence in order for alarm processing, excluding lastOccurrence for the first events batch and autoclear for the final events batch.